% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotator.R
\name{compareCelltypes}
\alias{compareCelltypes}
\title{Compare Cell Types Using Different Models}
\usage{
compareCelltypes(
  tissue,
  celltypes,
  marker,
  species = "human",
  model_list = NULL,
  output_file = NULL
)
}
\arguments{
\item{tissue}{Character string specifying the tissue type}

\item{celltypes}{Character vector of 2-4 cell types to compare}

\item{marker}{Character vector or string of marker genes}

\item{species}{Character string specifying the species (default: "human")}

\item{model_list}{Character vector of model names to use (optional)}
}
\value{
A list containing responses from different models
}
\description{
Compare Cell Types Using Different Models
}
