% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_settings.R
\name{.get_py_model_settings}
\alias{.get_py_model_settings}
\title{Get Python CASSIA Module}
\usage{
.get_py_model_settings()
}
\value{
Python CASSIA module (all functions available at package level)
}
\description{
Get Python CASSIA Module
}
\keyword{internal}
