% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotator.R
\name{.require_python}
\alias{.require_python}
\title{Check if Python environment is ready and stop with helpful message if not}
\usage{
.require_python(func_name = NULL)
}
\description{
Check if Python environment is ready and stop with helpful message if not
}
\keyword{internal}
