% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{listAvailableMarkers}
\alias{listAvailableMarkers}
\title{List Available Built-in Marker Sets}
\usage{
listAvailableMarkers()
}
\value{
A character vector containing names of available marker files
}
\description{
This function lists all available built-in marker sets in the CASSIA package.
}
\examples{
available_markers <- listAvailableMarkers()
print(available_markers)
}
