% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{loadExampleMarkers_subcluster}
\alias{loadExampleMarkers_subcluster}
\title{Load Example Subcluster Results}
\usage{
loadExampleMarkers_subcluster()
}
\value{
A data frame containing subcluster results
}
\description{
This function loads example subcluster results data that can be used to test CASSIA's
functionality.
}
\examples{
subcluster_data <- loadExampleMarkers_subcluster()
head(subcluster_data)
}
