% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotator.R
\name{runCASSIA_annottaionboost_additional_task}
\alias{runCASSIA_annottaionboost_additional_task}
\title{Generate Cell Type Analysis Report}
\usage{
runCASSIA_annottaionboost_additional_task(
  full_result_path,
  marker,
  cluster_name,
  major_cluster_info,
  output_name,
  num_iterations = 5,
  model = "anthropic/claude-3.5-sonnet",
  additional_task = ""
)
}
\arguments{
\item{full_result_path}{Path to the full results CSV file}

\item{cluster_name}{Name of the cluster to analyze}

\item{major_cluster_info}{General information about the dataset (e.g., "Human PBMC")}

\item{output_name}{Name of the output HTML file}

\item{num_iterations}{Number of iterations for marker analysis (default=5)}

\item{model}{Model to use for analysis (default="gpt-4o")}

\item{additional_task}{Additional task to perform}

\item{marker_path}{Path to the marker genes CSV file}
}
\value{
None
}
\description{
Generate Cell Type Analysis Report
}
