% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_settings.R
\name{.get_py_model_settings}
\alias{.get_py_model_settings}
\title{Get Python Model Settings Module}
\usage{
.get_py_model_settings()
}
\value{
Python model settings module
}
\description{
Get Python Model Settings Module
}
\keyword{internal}
